chore: move util to a separate file
